+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/configuration/Locations.java
@@ -19,6 +19,7 @@ public class Locations {
     private String config_integrator_package;
     private String config_sql_common_backend_package;
     private String config_sql_integration_backend_package;
+    private String config_sql_access_control_backend_package;
     public final String python_dir;
 
 
@@ -35,6 +36,7 @@ public class Locations {
         this.config_backend            = config.package_;
         this.config_sql_common_backend_package = config_backend + ".sql.common";
         this.config_sql_integration_backend_package = config_backend + ".sql.integration";
+        this.config_sql_access_control_backend_package = config_backend + ".sql.access_control";
     }
 
     public Locations(TemplatesProjectConfiguration configs, String cli_src_dir, String l2p_src_dir) {
@@ -79,7 +81,17 @@ public class Locations {
             case SQL_BEAN_ENACTOR3:
             case SQL_ENACTOR_CONFIGURATOR3:
             case SQL_COMPOSITE_ENACTOR_CONFIGURATOR3:
+            case BEAN_ENACTOR2_COMPOSITE_WP:
+            case BEAN_ENACTOR2_WP:
+            case QUERY_INVOKER2WP:
                 return config_sql_integration_backend_package;
+
+            case SQL_BEAN_ENACTOR4:
+            case SQL_ENACTOR_CONFIGURATOR4:
+            case SQL_COMPOSITE_BEAN_ENACTOR4:
+            case SQL_COMPOSITE_ENACTOR_CONFIGURATOR4:
+                return config_sql_access_control_backend_package;
+
             case LOGGER:
             case TEMPLATE_BUILDERS:
                 return logger_package;