+++ RELEASES.md
@@ -2,6 +2,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 
 `2.0.8-SNAPSHOT`
 * module-tools: prov-template-archetype and prov-toolbox-plugin
+* prov-template-compiler: record index with principal
 
 `2.0.7`
 * prov-jsonld: fixed context url